CR checkpoint
